refactor: simplify common:novel slice.
