--- lib/common/novel/data/data.dart
@@ -1,10 +1,0 @@
-// Routines.
-export 'routines/update_novel.dart';
-export 'routines/add_chapter_updates.dart';
-export 'routines/parse_novel_query.dart';
-
-// Mappers.
-export 'mappers/source_novel_into_companion.dart';
-export 'mappers/source_chapter_into_companion.dart';
-export 'mappers/source_volume_into_companion.dart';
-export 'mappers/source_metadata_into_companion.dart';
