--- lib/common/novel/domain/domain.dart
@@ -1,20 +1,0 @@
-// Entities.
-export 'entities/chapter_data.dart';
-export 'entities/novel_data.dart';
-export 'entities/partial_novel_data.dart';
-export 'entities/meta_entry_data.dart';
-export 'entities/volume_data.dart';
-
-// Failures.
-export 'failures/novel_not_found.dart';
-
-// Services.
-export 'services/fetch_chapter_content.dart';
-export 'services/fetch_novel.dart';
-export 'services/get_first_unread_chapter.dart';
-export 'services/get_novel_by_id.dart';
-export 'services/get_novel_by_url.dart';
-export 'services/get_novel_categories.dart';
-export 'services/get_novel_category_map.dart';
-export 'services/set_read_at.dart';
-export 'services/change_novel_categories.dart';
