+++ lib/common/novel/entities/chapter_data.dart
