+++ lib/common/novel/entities/meta_entry_data.dart
