+++ lib/common/novel/entities/partial_novel_data.dart
