+++ lib/common/novel/entities/volume_data.dart
