+++ lib/common/novel/failures/all_chapters_read.dart
