+++ lib/common/novel/failures/novel_not_found.dart
