+++ lib/common/novel/mappers/source_chapter_into_companion.dart
