+++ lib/common/novel/mappers/source_metadata_into_companion.dart
