+++ lib/common/novel/mappers/source_novel_into_companion.dart
