+++ lib/common/novel/novel.dart
@@ -1,3 +1,43 @@
-export 'data/data.dart';
-export 'domain/domain.dart';
-export 'presentation/presentation.dart';
+// Routines.
+export 'routines/update_novel.dart';
+export 'routines/add_chapter_updates.dart';
+export 'routines/parse_novel_query.dart';
+
+// Mappers.
+export 'mappers/source_novel_into_companion.dart';
+export 'mappers/source_chapter_into_companion.dart';
+export 'mappers/source_volume_into_companion.dart';
+export 'mappers/source_metadata_into_companion.dart';
+
+// Entities.
+export 'entities/chapter_data.dart';
+export 'entities/novel_data.dart';
+export 'entities/partial_novel_data.dart';
+export 'entities/meta_entry_data.dart';
+export 'entities/volume_data.dart';
+
+// Failures.
+export 'failures/novel_not_found.dart';
+
+// Services.
+export 'services/fetch_chapter_content.dart';
+export 'services/fetch_novel.dart';
+export 'services/get_first_unread_chapter.dart';
+export 'services/get_novel_by_id.dart';
+export 'services/get_novel_by_url.dart';
+export 'services/get_novel_categories.dart';
+export 'services/get_novel_category_map.dart';
+export 'services/set_read_at.dart';
+export 'services/change_novel_categories.dart';
+
+// Models.
+export 'models/novel_type.dart';
+
+// Providers.
+export 'providers/chapter_family.dart';
+export 'providers/chapter_list_family.dart';
+export 'providers/novel_family.dart';
+
+// Set categories
+export 'set_categories/provider/selected_categories_provider.dart';
+export 'set_categories/set_categories_dialog.dart';
