--- lib/common/novel/presentation/presentation.dart
@@ -1,11 +1,0 @@
-// Models.
-export 'models/novel_type.dart';
-
-// Providers.
-export 'providers/chapter_family.dart';
-export 'providers/chapter_list_family.dart';
-export 'providers/novel_family.dart';
-
-// Set categories
-export 'set_categories/provider/selected_categories_provider.dart';
-export 'set_categories/set_categories_dialog.dart';
