+++ lib/common/novel/services/change_novel_categories.dart
@@ -5,6 +5,8 @@ import 'package:nacht/core/core.dart';
 import 'package:nacht/database/database.dart';
 import 'package:drift/drift.dart';
 
+import '../entities/novel_data.dart';
+
 final changeNovelCategoriesProvider = Provider<ChangeNovelCategories>(
   (ref) => ChangeNovelCategories(
     database: ref.watch(databaseProvider),
