+++ lib/common/novel/services/get_first_unread_chapter.dart
@@ -1,10 +1,12 @@
 import 'package:dartz/dartz.dart';
 import 'package:drift/drift.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:nacht/common/common.dart';
-import 'package:nacht/common/novel/domain/failures/all_chapters_read.dart';
 import 'package:nacht/core/core.dart';
 import 'package:nacht/database/database.dart';
+
+import '../entities/chapter_data.dart';
+import '../entities/volume_data.dart';
+import '../failures/all_chapters_read.dart';
 
 final getFirstUnreadChapterProvider = Provider<GetFirstUnreadChapter>(
   (ref) => GetFirstUnreadChapter(
