+++ lib/common/novel/services/get_novel_by_id.dart
@@ -1,9 +1,11 @@
 import 'package:dartz/dartz.dart';
 import 'package:drift/drift.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:nacht/common/novel/novel.dart';
 import 'package:nacht/core/core.dart';
 import 'package:nacht/database/database.dart';
+
+import '../entities/novel_data.dart';
+import '../routines/parse_novel_query.dart';
 
 final getNovelByIdProvider = Provider<GetNovelById>(
   (ref) => GetNovelById(
