--- lib/common/selection/presentation/presentation.dart
@@ -1,5 +1,0 @@
-// Providers.
-export 'providers/selection_provider.dart';
-
-// Widgets.
-export 'widgets/sliver_selection_app_bar.dart';
