+++ lib/common/selection/selection.dart
@@ -1,1 +1,5 @@
-export 'presentation/presentation.dart';
+// Providers.
+export 'providers/selection_provider.dart';
+
+// Widgets.
+export 'widgets/sliver_selection_app_bar.dart';
